PassManager/requirements.txt
flask
mysql-connector
pycryptodome
numpy
redis
flask-wtf
more-itertools